==> Research/01-Data_Exploration.ipynb <==
import numpy as np
import pandas as pd

def read_data(path):
    return pd.read_csv(path)

path = "Data\job_descriptions.csv"

dataset = read_data(path = path)

dataset.head()

dataset.info()

dataset.describe()

dataset.describe(include = "O")
