==> Research/02-Data_preprocessing.ipynb <==
import pandas as pd 
import numpy as np 

## Importing the data 

file_path = 'Data\\job_descriptions.csv'
df = pd.read_csv(file_path)
df.head()

#### Identifying and handling missing value and duplicate value

missing_values = df.isnull().sum()
print("Missing Values:")
print(missing_values)
duplicate_rows = df.duplicated().sum()
print("\nDuplicate Rows:", duplicate_rows)

df = df.dropna(subset=['Company Profile']) #droping the column which has missing company profile (5478 Raws)

##### Droping Unnecessary column 

df = df.drop(columns=['Job Id', 'latitude', 'longitude', 'Contact Person', 'Contact', 'Job Portal', 'Benefits'])

df.info()

##### Extracting Min_experience and max_experience from experience and then droping the experience column 

# Defining a function to process the experience range
def extract_experience(experience_range):
    # Removing 'years' and spliting by 'to'
    parts = experience_range.replace(' Years', '').split(' to ')
    
    # Converting to integer
    min_experience = int(parts[0])
    max_experience = int(parts[1])
    
    return min_experience, max_experience

# Apply the function to extract min and max experience
df[['min_experience', 'max_experience']] = df['Experience'].apply(lambda x: pd.Series(extract_experience(x)))

# Display the updated DataFrame
df.head()

df = df.drop(columns='Experience')

# Calculating the average experience
df['average_experience'] = df[['min_experience', 'max_experience']].mean(axis=1)

##### Extracting min_salary and max_salary from salary_range and then droping the salary_range column 

def extract_salary(salary_range):
    # Removing '$' and 'K', then spliting by '-'
    parts_salary = salary_range.replace('$', '').replace('K', '').split('-')
    
    # Converting to integers and multiplying by 1000
    min_salary = int(parts_salary[0]) * 1000
    max_salary = int(parts_salary[1]) * 1000
    
    return min_salary, max_salary

# Applying the function to extract min and max salary
df[['min_salary', 'max_salary']] = df['Salary Range'].apply(lambda x: pd.Series(extract_salary(x)))

df = df.drop(columns='Salary Range')
df['average_salary'] = df[['min_salary', 'max_salary']].mean(axis=1)
df.head()

df['Job Posting Date'].info()

import pandas as pd
import re

irrelevant_words = [
    "knowledge", "and", "skills", "control", "including"
    "codes", "regulations", "cost", "monitor", "timelines",
    "assist", "in", "budgeting"
]

# Function to clean and extract skills
pattern = re.compile(r'\b(?:' + '|'.join(irrelevant_words) + r')\b', re.IGNORECASE)

# Function to clean and extract skills
def clean_skills(skills_string):
    # Remove brackets and their contents
    skills_string = re.sub(r'\[.*?\]|\(.*?\)|\{.*?\}', '', skills_string)
    
    # Clean the skills string by removing irrelevant words
    cleaned_skills = pattern.sub('', skills_string)
    
    # Split the cleaned string into a list of skills
    skill_list = re.split(r'\s*,\s*|\s+', cleaned_skills.strip())
    
    # Remove any empty strings from the list
    skill_list = [skill for skill in skill_list if skill]
    
    return skill_list

# Apply the function to the 'skills' column and generate the list of cleaned skills
df['skills'] = df['skills'].apply(clean_skills)

df['Job Posting Date'] = pd.to_datetime(df['Job Posting Date'])

df['Job Posting Date'].info()

df.info()

df.to_csv("Data/cleaned_df.csv", index= False)
