==> Research/04-Feature_engineering.ipynb <==
import numpy as np
import pandas as pd

import matplotlib.pyplot as plt
import seaborn as sns

dataset = pd.read_csv("Data/cleaned_df.csv")

dataset.info()

for col in dataset.columns:
    if dataset[col].dtype == "object":
        print(f"{col}: {len(dataset[col].value_counts().keys())}")

dataset["Qualifications"].unique()

bachelor = ["BCA", "BBA", "B.Tech", "B.Com", "BA"]
masters = ["MBA", "MCA", "M.Com", "M.Tech"]
phd = ["PhD"]

def map_qualification_to_level(qualifications):
    if qualifications in bachelor:
        return "Bachelor"
    elif qualifications in masters:
        return "Master"
    else:
        return "PhD"

dataset["Level"] = dataset["Qualifications"].apply(map_qualification_to_level)

dataset[["Level", "Qualifications"]].head()

bins = [0, 20000, 50000, 100000]
labels = ['small', 'medium', 'large']

dataset["Company_size"] = pd.cut(dataset["Company Size"], bins = bins, labels = labels, right = True, include_lowest = True)

dataset.info()

dataset.columns

dataset[['Qualifications', 'location', 'Country', 'Work Type', 'Company Size',
       'Job Posting Date', 'Preference', 'Job Title', 'Role',
       'Job Description', 'skills']].head()

drop_cols = ["location", "Country", "Job Posting Date", "Role", "Job Description", "Company", "Company Profile"]

dataset.drop(drop_cols, axis = 1, inplace = True)

dataset.info()

dataset.head()

# data = pd.get_dummies(dataset, columns = ["skills"])

dataset.shape

dataset.shape

# Normalize and tokenize the skills
dataset['skills'] = dataset['skills'].apply(lambda x: [skill.lower().strip() for skill in x.split()])

def load_glove_embeddings(file_path):
    embeddings_index = {}
    with open(file_path, 'r', encoding='utf-8') as f:
        for line in f:
            values = line.split()
            word = values[0]
            coefs = np.asarray(values[1:], dtype='float32')
            embeddings_index[word] = coefs
    return embeddings_index

glove_embeddings = load_glove_embeddings('Data/glove.6B.100d.txt')

def get_skill_embedding(skill, embeddings_index, embedding_dim=100):
    return embeddings_index.get(skill, np.zeros(embedding_dim))

# Create embeddings for each skill
dataset['skill_embeddings'] = dataset['skills'].apply(lambda skills: 
    [get_skill_embedding(skill, glove_embeddings) for skill in skills])

def aggregate_embeddings(skill_embeddings):
    if len(skill_embeddings) == 0:
        return np.zeros(len(skill_embeddings[0]))
    return np.mean(skill_embeddings, axis=0)

# Aggregate embeddings for each job posting
dataset['aggregated_skill_embeddings'] = dataset['skill_embeddings'].apply(aggregate_embeddings)

# Convert the aggregated skill embeddings into a matrix
embeddings_matrix = np.vstack(dataset['aggregated_skill_embeddings'].values)

# Display the shape of the embeddings matrix
print(embeddings_matrix.shape)

from sklearn.decomposition import PCA

pca = PCA(n_components=40)
reduced_data = pca.fit_transform(embeddings_matrix)

# Convert the reduced NumPy array to a DataFrame
reduced_df = pd.DataFrame(reduced_data, columns=[f'feature_{i+1}' for i in range(reduced_data.shape[1])])

# Concatenate the original DataFrame with the reduced DataFrame
merged_df = pd.concat([dataset, reduced_df], axis=1)

print(merged_df.shape)

merged_df.columns

dataset = merged_df.copy()

cols = ["Qualifications", "Work Type", "Company Size", "Preference", "Job Title",
        "min_experience","max_experience", "average_experience", "min_salary", 
        "max_salary","average_salary", "Level", "Company_size",
        "feature_1", "feature_2", "feature_3",
        "feature_4", "feature_5", "feature_6", "feature_7", "feature_8",
        "feature_9", "feature_10", "feature_11", "feature_12", "feature_13",
        "feature_14", "feature_15", "feature_16", "feature_17", "feature_18",
        "feature_19", "feature_20", "feature_21", "feature_22", "feature_23",
        "feature_24", "feature_25", "feature_26", "feature_27", "feature_28",
        "feature_29", "feature_30", "feature_31", "feature_32", "feature_33",
        "feature_34", "feature_35", "feature_36", "feature_37", "feature_38",
        "feature_39", "feature_40"]

dataset = dataset[cols]

dataset.head()

import pandas as pd
from sklearn.preprocessing import LabelEncoder

# Apply one-hot encoding
one_hot = pd.get_dummies(dataset[["Qualifications", "Work Type", "Preference"]])

# Apply label encoding to other categorical columns
label_encoders = {}
for col in ["Company Size", "Level", "Company_size"]:
    le = LabelEncoder()
    dataset[col] = le.fit_transform(dataset[col].astype(str))
    label_encoders[col] = le

# Combine one-hot encoded columns with the label encoded columns
df_encoded = pd.concat([dataset.drop(columns=["Qualifications", "Work Type", "Preference"]), one_hot], axis=1)

print(df_encoded.shape)

df_encoded.to_csv("Data/data_after_feature_eng.csv", index = False)
