==> Research/05-model_selection1.ipynb <==
## Creating a deep learning model for salary prediction 

import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_absolute_error, mean_squared_error, r2_score
from sklearn.preprocessing import StandardScaler, OneHotEncoder
from sklearn.compose import ColumnTransformer
from sklearn.pipeline import Pipeline
import tensorflow as tf
from tensorflow.keras import layers, models

# Importing the data 
path = "Data/data_after_feature_eng.csv"
df = pd.read_csv(path)
df.head()

df.info()

# Drop unnecessary columns
df = df.drop(columns=['min_salary', 'max_salary', 'Company_size'])

columns_to_drop = [
    'Work Type_Contract',
    'Work Type_Full-Time',
    'Work Type_Intern',
    'Work Type_Part-Time',
    'Work Type_Temporary'
]

# Drop the specified columns
df = df.drop(columns=columns_to_drop)

### Handling Categorical Variables

# Job Title is the only categorical feature that needs one-hot encoding
categorical_features = ['Job Title']
numeric_features = df.columns[df.columns.str.startswith('feature_')].tolist() + [
    'min_experience', 'max_experience', 'average_experience', 'Company Size', 'Level'
]

# Handling Numerical Variables and Creating Preprocessing Pipeline
preprocessor = ColumnTransformer(
    transformers=[
        ('num', StandardScaler(), numeric_features),
        ('cat', OneHotEncoder(), categorical_features)
    ])

### Splitting Data

X = df.drop(columns=['average_salary'])
y = df['average_salary']

X_preprocessed = preprocessor.fit_transform(X)

# Splitting Data
X_train, X_test, y_train, y_test = train_test_split(X_preprocessed, y, test_size=0.2)

### Model Building

# Build the neural network model
model = models.Sequential([
    layers.Dense(256, activation='relu', input_shape=(X_train.shape[1],)),
    layers.Dense(512, activation='relu'),
    layers.Dense(1024, activation='relu'),
    layers.Dense(1024, activation='relu'),
    layers.Dense(1)  # Single output for regression
])

# Compile the model
model.compile(optimizer='adam', loss='mse', metrics=['mae'])

# Train the model
history = model.fit(X_train, y_train, epochs=5, validation_split=0.2, batch_size=1024)

# Evaluate the model
test_loss, test_mae = model.evaluate(X_test, y_test)
print(f"Test MAE: {test_mae}")

# Predict on the test set
y_pred = model.predict(X_test)

# Calculate MSE
mse = mean_squared_error(y_test, y_pred)
print(f"Mean Squared Error (MSE): {mse}")

# Calculate RMSE
rmse = np.sqrt(mse)
print(f"Root Mean Squared Error (RMSE): {rmse}")

# Calculate R-Squared
r2 = r2_score(y_test, y_pred)
print(f"R-Squared (R²): {r2}")

# Build the neural network model_2
model = models.Sequential([
    layers.Dense(256, activation='relu', input_shape=(X_train.shape[1],)),
    layers.Dense(512, activation='relu'),
    layers.Dense(512, activation='relu'),
    layers.Dense(512, activation='relu'),
    layers.Dense(1024, activation='relu'),
    layers.Dense(1024, activation='relu'),
    layers.Dense(1)
])

history = model.fit(X_train, y_train, epochs=5, validation_split=0.2, batch_size=1024)

model.compile(optimizer='adam', loss='mse', metrics=['mae'])

# Evaluate the model
test_loss, test_mae = model.evaluate(X_test, y_test)
print(f"Test MAE: {test_mae}")

# Predict on the test set
y_pred = model.predict(X_test)

# Calculate MSE
mse = mean_squared_error(y_test, y_pred)
print(f"Mean Squared Error (MSE): {mse}")

# Calculate RMSE
rmse = np.sqrt(mse)
print(f"Root Mean Squared Error (RMSE): {rmse}")

# Calculate R-Squared
r2 = r2_score(y_test, y_pred)
print(f"R-Squared (R²): {r2}")

# Build the neural network model_2
model = models.Sequential([
    layers.Dense(256, activation='relu', input_shape=(X_train.shape[1],)),
    layers.Dense(512, activation='relu'),
    layers.Dense(1) 
])

history = model.fit(X_train, y_train, epochs=7, validation_split=0.2, batch_size=1024)

model.compile(optimizer='adam', loss='mse', metrics=['mae'])

# Evaluate the model
test_loss, test_mae = model.evaluate(X_test, y_test)
print(f"Test MAE: {test_mae}")

# Predict on the test set
y_pred = model.predict(X_test)

# Calculate MSE
mse = mean_squared_error(y_test, y_pred)
print(f"Mean Squared Error (MSE): {mse}")

# Calculate RMSE
rmse = np.sqrt(mse)
print(f"Root Mean Squared Error (RMSE): {rmse}")

# Calculate R-Squared
r2 = r2_score(y_test, y_pred)
print(f"R-Squared (R²): {r2}")

import math 

# Custom quadratic activation function
def quadratic_activation(x):
    return tf.math.square(x)

# Register the custom activation function so that it can be used with Keras models
tf.keras.utils.get_custom_objects().update({'quadratic_activation': layers.Activation(quadratic_activation)})

# Build the neural network model with a quadratic output layer
model = models.Sequential([
    layers.Dense(256, activation='relu', input_shape=(X_train.shape[1],)),
    layers.Dense(512, activation='relu'),
    layers.Dense(1, activation=quadratic_activation)  # Custom quadratic activation function
])

# Compile the model
model.compile(optimizer='adam', loss='mse', metrics=['mae'])

# Print the model summary
model.summary()


history = model.fit(X_train, y_train, epochs=7, validation_split=0.2, batch_size=2048)

# Evaluate the model
test_loss, test_mae = model.evaluate(X_test, y_test)
print(f"Test MAE: {test_mae}")

# Predict on the test set
y_pred = model.predict(X_test)

# Calculate MSE
mse = mean_squared_error(y_test, y_pred)
print(f"Mean Squared Error (MSE): {mse}")

# Calculate RMSE
rmse = np.sqrt(mse)
print(f"Root Mean Squared Error (RMSE): {rmse}")

# Calculate R-Squared
r2 = r2_score(y_test, y_pred)
print(f"R-Squared (R²): {r2}")

import math 

# Custom quadratic activation function
def quadratic_activation(x):
    return tf.math.square(x)

# Register the custom activation function so that it can be used with Keras models
tf.keras.utils.get_custom_objects().update({'quadratic_activation': layers.Activation(quadratic_activation)})

# Build the neural network model with a quadratic output layer
model = models.Sequential([
    layers.Dense(256, activation='relu', input_shape=(X_train.shape[1],)),
    layers.Dense(1, activation=quadratic_activation)  # Custom quadratic activation function
])

# Compile the model
model.compile(optimizer='adam', loss='mse', metrics=['mae'])

# Print the model summary
model.summary()


history = model.fit(X_train, y_train, epochs=7, validation_split=0.2, batch_size=1024)

# Evaluate the model
test_loss, test_mae = model.evaluate(X_test, y_test)
print(f"Test MAE: {test_mae}")

# Predict on the test set
y_pred = model.predict(X_test)

# Calculate MSE
mse = mean_squared_error(y_test, y_pred)
print(f"Mean Squared Error (MSE): {mse}")

# Calculate RMSE
rmse = np.sqrt(mse)
print(f"Root Mean Squared Error (RMSE): {rmse}")

# Calculate R-Squared
r2 = r2_score(y_test, y_pred)
print(f"R-Squared (R²): {r2}")


from tensorflow.keras.optimizers import Adam

# Custom quadratic activation function
def quadratic_activation(x):
    return tf.math.square(x)

# Register the custom activation function so that it can be used with Keras models
tf.keras.utils.get_custom_objects().update({'quadratic_activation': layers.Activation(quadratic_activation)})

# Build the neural network model with a quadratic output layer
model = models.Sequential([
    layers.Dense(256, activation='relu', input_shape=(X_train.shape[1],)),
    layers.Dense(512, activation='relu'),
    layers.Dense(1024, activation='relu'),
    layers.Dense(1024, activation='relu'),
    layers.Dense(1, activation=quadratic_activation)  # Custom quadratic activation function
])

# Compile the model


# Example: Reduce learning rate
model.compile(optimizer=Adam(learning_rate=0.0001), loss='mse', metrics=['mae'])


# Print the model summary
model.summary()


# Train for more epochs
history = model.fit(X_train, y_train, epochs=5, validation_split=0.2, batch_size=2048)


test_loss, test_mae = model.evaluate(X_test, y_test)
print(f"Test MAE: {test_mae}")

# Predict on the test set
y_pred = model.predict(X_test)

# Calculate MSE
mse = mean_squared_error(y_test, y_pred)
print(f"Mean Squared Error (MSE): {mse}")

# Calculate RMSE
rmse = np.sqrt(mse)
print(f"Root Mean Squared Error (RMSE): {rmse}")

# Calculate R-Squared
r2 = r2_score(y_test, y_pred)
print(f"R-Squared (R²): {r2}")

import pickle

import pickle

# Assuming 'model' is your trained model
with open('Data/model_salary.pkl', 'wb') as file:
    pickle.dump(model, file)

df.info()

import pandas as pd
import numpy as np
import statsmodels.api as sm
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LinearRegression
from sklearn.metrics import mean_squared_error, r2_score

X = df.drop('average_salary', axis=1)  
y = df['average_salary'] 

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

import pandas as pd

# Assuming 'df' is your original dataframe
job_title_encoded = pd.get_dummies(df['Job Title'], prefix='Job_Title')

# Display the first few rows of the encoded 'Job Title' columns
print(job_title_encoded.head())

# If you want to combine this with your original dataframe
df_encoded = pd.concat([df, job_title_encoded], axis=1)

# Drop the original 'Job Title' column, as it's now encoded
df_encoded.drop('Job Title', axis=1, inplace=True)

# Display the first few rows of the updated dataframe
print(df_encoded.head())


X = df_encoded.drop('average_salary', axis=1)  
y = df_encoded['average_salary'] 

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=1)

from sklearn.linear_model import LinearRegression

# Initialize the linear regression model
model = LinearRegression()

# Fit the model to the training data
model.fit(X_train, y_train)

from sklearn.metrics import r2_score, mean_absolute_error, mean_squared_error
import numpy as np

# Make predictions on the training or test data
y_pred = model.predict(X_test)

# Calculate R² score
r2 = r2_score(y_test, y_pred)

# Calculate Mean Absolute Error (MAE)
mae = mean_absolute_error(y_test, y_pred)

# Calculate Mean Squared Error (MSE)
mse = mean_squared_error(y_test, y_pred)

# Calculate Root Mean Squared Error (RMSE)
rmse = np.sqrt(mse)

print(f'R² score: {r2}')
print(f'Mean Absolute Error (MAE): {mae}')
print(f'Mean Squared Error (MSE): {mse}')
print(f'Root Mean Squared Error (RMSE): {rmse}')


import numpy as np
from lime.lime_tabular import LimeTabularExplainer
from sklearn.linear_model import LinearRegression

# Initialize LIME explainer for a regression model
explainer = LimeTabularExplainer(
    training_data=np.array(X_train),
    feature_names=X_train.columns,  # assuming X_train is a DataFrame
    mode='regression'
)

# Explain a prediction
i = 0  # First instance in the test set
exp = explainer.explain_instance(
    data_row=X_test.iloc[i], 
    predict_fn=model.predict
)

# Get the most important feature
importance_list = exp.as_list()
most_important_feature = importance_list[0]

# Output the most important feature
print("Most important feature for this prediction:")
print(most_important_feature)
